draft version of the chatroom server update: have a message and a member model for further database preperation. Setup the chatroom server framework with all the to_json methods ready as we want to transform json instead of raw string data. I also refactored the code a little bit so we have a clearer structure
