+++ src/servers/Server2.java
@@ -1,4 +1,4 @@
-package model;
+package servers;
 
 import java.io.BufferedReader;
 import java.io.IOException;
